refactor(app): restructure types and extract guards into utils (#105)
Closes: #105
