+++ src/lib/components/WeatherIcons/WeatherIcons.types.ts
@@ -1,0 +1,6 @@
+import components from './';
+
+/**
+ * Type representing the available weather icon components.
+ */
+export type WeatherIconComponent = keyof typeof components;
