+++ src/lib/stores/health.types.ts
@@ -1,0 +1,8 @@
+import type { NightScoutData } from '../types';
+
+/**
+ * Type representing health data.
+ */
+export type HealthData = {
+  nightScout?: false | NightScoutData;
+};
