+++ src/lib/stores/index.ts
@@ -3,5 +3,5 @@ export { streams } from './streams.svelte';
 export { createWidget } from './widget.svelte';
 export { createHomeState, homeState } from './home.svelte';
 export { createHealthState, healthState } from './health.svelte';
-export { timeState, uiState } from './client.svelte';
+export { timeState, uiState } from './ui.svelte';
 export { musicState } from './music.svelte';
