+++ src/lib/stores/solar.types.ts
@@ -1,0 +1,8 @@
+/**
+ * Type representing solar data with sunrise and sunset times.
+ */
+export type SolarData = {
+  sunrise: Date[];
+  sunset: Date[];
+  isDay: number;
+};
