+++ src/lib/stores/ui.types.ts
@@ -1,0 +1,8 @@
+export type UIState = {
+  modalState: ModalState;
+};
+
+export type ModalState = {
+  isActive: boolean;
+  component: 'music' | 'nightscout' | 'weather' | undefined;
+};
