+++ src/lib/stores/weatherLeg.ts
@@ -1,6 +1,6 @@
-import type { CurrentWeather } from '$types';
+import type { CurrentWeatherData } from '$types';
 import { writable, type Writable } from 'svelte/store';
 
-const weather: Writable<CurrentWeather> = writable();
+const weather: Writable<CurrentWeatherData> = writable();
 
 export default weather;
