+++ src/lib/utils/guards/index.ts
@@ -1,0 +1,10 @@
+export * from './isLocationSettings';
+export * from './isCurrentWeatherData';
+export * from './isDailyWeatherData';
+export * from './isHueActionType';
+export * from './isLocationSettings';
+export * from './isNightScoutData';
+export * from './isSeptaNextToArriveData';
+export * from './isWeatherData';
+export * from './isTypeOfWidget';
+export * from './isWidgetSettings';
