+++ src/lib/utils/guards/isHueActionType.ts
@@ -1,0 +1,5 @@
+import type { HueActionType } from '$types';
+
+export const isHueActionType = (value: string): value is HueActionType => {
+  return value === 'lights' || value === 'groups';
+};
