+++ src/lib/utils/guards/isTypeOfWidget.ts
@@ -1,0 +1,5 @@
+import { TypeOfWidget } from '$types';
+
+export const isTypeOfWidget = (type: string): type is TypeOfWidget => {
+  return Object.values(TypeOfWidget).includes(type as TypeOfWidget);
+};
