+++ src/lib/utils/nightscout.ts
@@ -1,8 +1,5 @@
-import {
-  isNightScoutData,
-  type ChartSeriesGlucose,
-  type NightScoutData,
-} from '$types';
+import { type ChartSeriesGlucose, type NightScoutData } from '$types';
+import { isNightScoutData } from '$guards';
 
 // Updated type for mapNightScoutDirectionIcon to accept NightScoutData | false
 export const mapNightScoutDirectionIcon = (
