+++ src/lib/widgets/Hue/hue.types.ts
@@ -1,5 +1,1 @@
 export type HueActionType = 'lights' | 'groups';
-
-export const isHueActionType = (value: string): value is HueActionType => {
-  return value === 'lights' || value === 'groups';
-};
