+++ src/lib/widgets/Septa/NextToArrive/NextToArrive.svelte
@@ -1,9 +1,6 @@
 <script lang="ts">
-  import {
-    isSeptaNextToArriveDataArray,
-    TypeOfWidget,
-    type SeptaNextToArriveData,
-  } from '$types';
+  import { TypeOfWidget, type SeptaNextToArriveData } from '$types';
+  import { isSeptaNextToArriveDataArray } from '$guards';
   import { createWidget } from '$stores';
   import { generateID, toCamelCase } from '$utils/strings';
   import Icon from '@iconify/svelte';
