+++ src/lib/widgets/Septa/Septa.types.ts
@@ -1,0 +1,11 @@
+/**
+ * Type representing SEPTA next-to-arrive data.
+ */
+export type SeptaNextToArriveData = {
+  orig_train: string;
+  orig_line: string;
+  orig_departure_time: string;
+  orig_arrival_time: string;
+  orig_delay: string;
+  isdirect: string;
+};
