+++ svelte.config.js
@@ -24,7 +24,8 @@ const config = {
       $layouts: 'src/routes/(layouts)',
       $assets: 'static',
       $root: 'src',
-      $types: 'src/lib/types.ts',
+      $types: 'src/lib/types',
+      $guards: 'src/lib/utils/guards',
       $stores: 'src/lib/stores',
     },
   },
